requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
groq
chromadb
langchain-text-splitters
langchain-community
sentence-transformers
python-dotenv
